+++ src/main/java/com/strangeone101/pixeltweaks/TweaksConfig.java
@@ -3,7 +3,6 @@ package com.strangeone101.pixeltweaks;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.config.ModConfig;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 public class TweaksConfig {
 
@@ -14,6 +13,7 @@ public class TweaksConfig {
     public static ForgeConfigSpec.ConfigValue<Integer> hypertrainLevel;
     public static ForgeConfigSpec.ConfigValue<Boolean> randomlyDropRipeApricorns;
     public static ForgeConfigSpec.ConfigValue<Boolean> foxesLoveBerries;
+    public static ForgeConfigSpec.ConfigValue<Boolean> antiTrampleCrops;
     public static ForgeConfigSpec.ConfigValue<Double> catchMultiplier;
     public static ForgeConfigSpec.ConfigValue<Double> legendaryCatchMultiplier;
     public static ForgeConfigSpec.ConfigValue<Boolean> includeMythicals;
@@ -21,15 +21,19 @@ public class TweaksConfig {
     //Client
     public static ForgeConfigSpec.ConfigValue<Double> shinySparkleRange;
     public static ForgeConfigSpec.ConfigValue<Double> shinySparkleVolume;
+    public static ForgeConfigSpec.ConfigValue<Integer> autoWrapLoreLength;
 
     public TweaksConfig() {
         BUILDER.comment("All features that require being on the client").push("Client Features");;
 
         shinySparkleRange = BUILDER.comment("The range in which shinies will sparkle. Set to -1 to disable. Default: 20.0")
-                .define("shinySparkleRange", 20.0);
+                .defineInRange("shinySparkleRange", 25.0, -1.0F, 500.0F);
 
         shinySparkleVolume = BUILDER.comment("The volume of the shiny sparkle sound. Set to 0 to disable. Default and maximum: 1.0")
-                .define("shinySparkleVolume", 1.0);
+                .defineInRange("shinySparkleVolume", 1.0, 0F, 1.0F);
+
+        autoWrapLoreLength = BUILDER.comment("The length of each line of lore in the item tooltip. Set to 0 to disable. Default: 60")
+                .define("autoWrapLoreLength", 60);
 
         BUILDER.pop().comment("All features that require being on the server").push("Server Features");
 
@@ -37,13 +41,16 @@ public class TweaksConfig {
                 .define("healersDropThemselves", true);
 
         hypertrainLevel = BUILDER.comment("The level a pokemon can begin to be hypertrained. Default: 50 (PixelTweaks), 100 (Pixelmon)")
-                .define("hypertrainLevel", 50);
+                .defineInRange("hypertrainLevel", 50, 1, 100);
 
         randomlyDropRipeApricorns = BUILDER.comment("If apricorns should randomly drop when they are ripe. Default: true")
                 .define("randomlyDropRipeApricorns", true);
 
         foxesLoveBerries = BUILDER.comment("If fox pokemon should become immune to vanilla berry bushes & heal when right clicked with berries. Default: true")
                 .define("foxesLoveBerries", true);
+
+        antiTrampleCrops = BUILDER.comment("If pokemon should be prevented from trampling crops. Default: true")
+                .define("antiTrampleCrops", true);
 
         BUILDER.pop();
         /*BUILDER.comment("All catch related features").push("Server Features - Catching");